+++ packages/moonstone/ContextualPopupDecorator/ContextualPopupDecorator.js
@@ -354,6 +354,7 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		spotActivator = (activator) => {
+			Spotlight.setActiveContainer(this.state.containerId);
 			if (!Spotlight.focus(activator)) {
 				Spotlight.focus();
 			}
@@ -362,7 +363,16 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		spotPopupContent = () => {
 			const {containerId} = this.state;
 			if (!Spotlight.focus(containerId)) {
-				Spotlight.setActiveContainer(containerId);
+				const current = Spotlight.getCurrent();
+
+				// In cases where the container contains no spottable controls or we're in pointer-mode, focus
+				// cannot inherently set the active container or blur the active control, so we must do that
+				// here.
+				if (current) {
+					current.blur();
+					Spotlight.setPointerMode(false);
+				}
+				Spotlight.setActiveContainer(containerId, containerId);
 			}
 		}